+++ src/config/parse.rs
@@ -102,6 +102,10 @@ pub(super) fn get_report_uri(args: &ArgMatches) -> Option<String> {
     args.value_of("report-uri").map(ToString::to_string)
 }
 
+pub(super) fn get_profile(args: &ArgMatches) -> Option<String> {
+    args.value_of("profile").map(ToString::to_string)
+}
+
 pub(super) fn get_outputs(args: &ArgMatches) -> Vec<OutputFile> {
     values_t!(args.values_of("out"), OutputFile).unwrap_or_else(|_| vec![])
 }
@@ -111,7 +115,29 @@ pub(super) fn get_output_directory(args: &ArgMatches) -> Option<PathBuf> {
 }
 
 pub(super) fn get_run_types(args: &ArgMatches) -> Vec<RunType> {
-    values_t!(args.values_of("run-types"), RunType).unwrap_or_else(|_| vec![RunType::Tests])
+    let mut res = values_t!(args.values_of("run-types"), RunType).unwrap_or_else(|_| vec![]);
+    if args.is_present("lib") && !res.contains(&RunType::Lib) {
+        res.push(RunType::Lib);
+    }
+    if args.is_present("all-targets") && !res.contains(&RunType::AllTargets) {
+        res.push(RunType::AllTargets);
+    }
+    if args.is_present("benches") && !res.contains(&RunType::Benchmarks) {
+        res.push(RunType::Benchmarks);
+    }
+    if args.is_present("bins") && !res.contains(&RunType::Bins) {
+        res.push(RunType::Bins);
+    }
+    if args.is_present("examples") && !res.contains(&RunType::Examples) {
+        res.push(RunType::Examples);
+    }
+    if args.is_present("doc") && !res.contains(&RunType::Doctests) {
+        res.push(RunType::Doctests);
+    }
+    if args.is_present("tests") && !res.contains(&RunType::Tests) {
+        res.push(RunType::Tests);
+    }
+    res
 }
 
 pub(super) fn get_excluded(args: &ArgMatches) -> Vec<Regex> {
@@ -120,7 +146,6 @@ pub(super) fn get_excluded(args: &ArgMatches) -> Vec<Regex> {
 
 pub(super) fn regexes_from_excluded(strs: &[String]) -> Vec<Regex> {
     let mut files = vec![];
-
     for temp_str in strs {
         let s = &temp_str.replace(".", r"\.").replace("*", ".*");